// types/testimonials.ts
export type TestimonialSliderSlide = {
  id: string;
  descriptionLead: string;
  descriptionSpan: string;
  photoSrc: string;
  photoAlt: string;
  name: string;
  rolePrefix: string;
  companyName: string;
};

export type TestimonialStickyCard = {
  id: string;
  descriptionLead: string;
  descriptionSpan: string;
  photoSrc: string;
  photoAlt: string;
  largeSrc?: string;
  largeAlt?: string;
  name: string;
  rolePrefix: string;
  companyName: string;
};